move strip_dict_params to plotly/test/utils.py
