+++ plotly/tests/test_core/test_graph_objs/test_update.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from unittest import skip
 
 from plotly.graph_objs import Data, Figure, Layout, Line, Scatter, XAxis
-
+from plotly.tests.utils import strip_dict_params
 
 def test_update_dict():
     title = 'this'
@@ -16,11 +16,14 @@ def test_update_dict():
 def test_update_list():
     trace1 = Scatter(x=[1, 2, 3], y=[2, 1, 2])
     trace2 = Scatter(x=[1, 2, 3], y=[3, 2, 1])
-    data = Data([trace1, trace2])
+    fig = Figure([trace1, trace2])
     update = dict(x=[2, 3, 4], y=[1, 2, 3])
-    data.update(update)
-    assert data[0] == Scatter(x=[2, 3, 4], y=[1, 2, 3])
-    assert data[1] == Scatter(x=[2, 3, 4], y=[1, 2, 3])
+    fig.data[0].update({'x': [2, 3, 4], 'y': [1, 2, 3]})
+    fig.data[1].update({'x': [2, 3, 4], 'y': [1, 2, 3]})
+
+    d1, d2 = strip_dict_params()
+    assert fig.data[0] == Scatter(x=[2, 3, 4], y=[1, 2, 3])
+    assert fig.data[1] == Scatter(x=[2, 3, 4], y=[1, 2, 3])
 
 
 def test_update_dict_empty():
